# perf: Optimize local attention memory allocation in FlashAttentionBackend