<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation
The current implementation allocates maximum sized buffers for local attention metadata during CUDA graph capture, regardless of the actual requirements of the batch being processed. This leads to significant memory overhead and reduced performance.

For instance, with Llama-4-Maverick-17B-128E-Instruct-FP8 on 8×H100 GPUs, we can serve a maximum of 535k tokens. If we set `context_len` to 524288 and attn_chunk_size to 8192, our theoretical `max_virtual_batches` would be 64 × 160 (where 64 is the maximum number of chunks per sequence and 160 is the maximum batch size). It is even bigger for `local_block_table` tensor.

However, it's practically impossible to have 160 concurrent requests each with 524k tokens, as this would far exceed our total token budget of 535k. Without optimized buffer allocation, we would always be sending 65 × 160 = 10240 shape tensors for local_attn_metadata to the CUDA graph, which wastes significant GPU memory and reduces overall inference throughput. For instance, if there are 160 requests, each has length less than 8192, the actual shape of `local_query_start_loc` would only be 160, but now we are sending a 10240 shape. As a result, it even takes 10 seconds for the server to finish the warmup `/generate` request. 

<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

## Modifications
Added a new method `_update_local_attn_metadata_for_capture` that:
- Calculates the precise dimensions needed for local attention metadata based on the actual batch being processed
- Creates optimized tensor views with exactly the required sizes

<!-- Describe the changes made in this PR. -->

## Benchmark

```
# Server: 8xH100
python3 -m sglang.launch_server --model meta-llama/Llama-4-Maverick-17B-128E-Instruct-FP8 --port=8080 --tp-size=8 --context-length=524288 --chat-template=llama-4
```

```
# benchmark
python3 -m sglang.bench_serving --backend sglang --dataset-name random --num-prompts 3000 --random-input 1000 --random-output 1000 --random-range-ratio 1.0 --max-concurrency 64 --port 8080
```

Before: `main` branch
```
Output token throughput (tok/s):         3136.87 
```

After: current branch
```
Output token throughput (tok/s):         4165.05
```

vllm `3164` from blog

### Evaluation

```
➜  sglang git:(chang/opt-local-attn) ✗ python3 -m sglang.eval.loogle_eval --api-url=http://127.0.0.1:8080/v1
Running benchmark: 100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1101/1101 [00:00<00:00, 85129.39it/s]
Some weights of RobertaModel were not initialized from the model checkpoint at roberta-large and are newly initialized: ['pooler.dense.bias', 'pooler.dense.weight']
You should probably TRAIN this model on a down-stream task to be able to use it for predictions and inference.
Loading responses: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1101/1101 [00:00<00:00, 5616.48it/s]
Scoring batches: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 18/18 [02:12<00:00,  7.36s/it]
Average BERTScore (F1): 84.38%
```

## Checklist

- [x] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [x] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [x] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.